+++ src/django_htmx/http.py
@@ -1,4 +1,7 @@
+from typing import Any
+
 from django.http import HttpResponse
+from django.http.response import HttpResponseRedirectBase
 
 HTMX_STOP_POLLING = 286
 
@@ -9,3 +12,12 @@ class HttpResponseStopPolling(HttpResponse):
     def __init__(self, *args: object, **kwargs: object) -> None:
         super().__init__(*args, **kwargs)
         self._reason_phrase = "Stop Polling"
+
+
+class HttpResponseClientRedirect(HttpResponseRedirectBase):
+    status_code = 200
+
+    def __init__(self, redirect_to: str, *args: Any, **kwargs: Any) -> None:
+        super().__init__(redirect_to, *args, **kwargs)
+        self["HX-Redirect"] = self["Location"]
+        del self["Location"]